Add Unit tests for AccountService, AuthService and EventService

+++ Software/Zvonko/BusinessLogicLayer/EventService.cs
@@ -11,6 +11,7 @@ namespace BusinessLogicLayer {
         
     public bool AddEvent(Event newEvent)
     {
+        if (newEvent == null) return false;
         using (var repo = new EventRepository())
         {
             int affectedRows = repo.Add(newEvent, true);
@@ -26,17 +27,18 @@ namespace BusinessLogicLayer {
             }
     }
 
-        public IEnumerable<Event> GetAllEventsAndRecordings() {
-            using (var repo = new EventRepository()) {
-                return repo.GetRecordingsAndEvents().ToList();
-            }
+    public IEnumerable<Event> GetAllEventsAndRecordings() {
+        using (var repo = new EventRepository()) {
+            return repo.GetRecordingsAndEvents().ToList();
         }
-        public bool RemoveEvent(Event eventToRemove) {
-            using (var repo = new EventRepository()) {
-                int affectedRows = repo.Remove(eventToRemove, true);
-                return affectedRows > 0;
-            }
+    }
+    public bool RemoveEvent(Event eventToRemove) {
+        if (eventToRemove == null) return false;
+        using (var repo = new EventRepository()) {
+            int affectedRows = repo.Remove(eventToRemove, true);
+            return affectedRows > 0;
         }
+    }
 
     }
 }
